% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_hello.R
\name{simple_hello}
\alias{simple_hello}
\title{simple hello Function}
\usage{
simple_hello()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
simple_hello()
}
\keyword{cats}
